move phase1 functions to package
